Moved widgets for lists and recipes into subfolders
