+++ lib/widgets/lists/shopping_list_entry.dart
