+++ lib/widgets/lists/shopping_list_item_entry.dart
