+++ lib/widgets/recipes/recipes_list_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:listeverywhere_app/models/recipe_model.dart';
-import 'package:listeverywhere_app/widgets/recipe_entry.dart';
+import 'package:listeverywhere_app/widgets/recipes/recipe_entry.dart';
 
 class RecipeListView extends StatelessWidget {
   const RecipeListView({
